requirements.txt
streamlit
scikit-learn
pandas
numpy
xgboost
joblib
matplotlib 
seaborn
fastapi==0.115.2
uvicorn[standard]==0.30.0
pydantic==2.9.2
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
mlflow==2.16.0
PyYAML==6.0.2
structlog==24.1.0
joblib==1.4.2
prometheus-fastapi-instrumentator==6.1.0
openpyxl==3.1.5
python-multipart==0.0.9